requirements.txt
numpy
pandas
tensorflow
scipy
keras
Pillow
scikit-image
matplotlib
argparse
opencv-python
imutils
scikit-learn
opencv-utils